// app/main/enums/messageboard.ts
export enum MessageboardStatus {
  DEACTIVATED = 0,
  ACTIVATED = 1,
}

export enum MessagePriority {
  UNPINNED = 0,
  PINNED = 1,
}

export enum MessageStatus {
  DELETED = 0,
  ACTIVE = 1,
  OMITTED = 2,
}

export enum MessageTone {
  DISABLED = 0,
  ENABLED = 1,
}

export enum MessageType {
  NOT_IMPORTANT = 0,
  IMPORTANT = 1,
}